Use komodo-test for PRs

+++ ci/jenkins/testkomodo.sh
@@ -1,16 +1,19 @@
 #!/bin/bash
 set -xe
+PROJECT=libres
 source $KOMODO_ROOT/$RELEASE_NAME/enable
 source $DEVTOOL/enable
 
 # find and check out the code that was used to build libres for this komodo relase
-echo "checkout tag from komodo"
-EV=$(cat $KOMODO_ROOT/$RELEASE_NAME/$RELEASE_NAME | grep "libres:" -A2 | grep "version:")
-EV=($EV)    # split the string "version: vX.X.X"
-EV=${EV[1]} # extract the version
-EV=${EV%"+py3"}
-echo "Using libres version $EV"
-git checkout $EV
+if [[ -z "${sha1// }" ]]; then
+    # this is not a PR build, the komodo everest verison is checked out
+    EV=$(cat ${RELEASE_PATH}/${RELEASE_NAME} | grep "${PROJECT}:" -A2 | grep "version:")
+    EV=($EV)    # split the string "version: vX.X.X"
+    EV=${EV[1]} # extract the version
+    EV=${EV%"+py3"}
+    echo "Using ${PROJECT} version ${EV}"
+    $GIT checkout $EV
+fi
 
 GCC_VERSION=7.3.0 CMAKE_VERSION=3.10.2 source $SDPSOFT/env.sh
 echo "building c tests"
@@ -26,8 +29,9 @@ cmake .. -DBUILD_TESTS=ON\
 -DCMAKE_CXX_FLAGS='-Wno-unused-result'
 make -j 12
 #removing built libs in order to ensure we are using libs from komodo
-rm -r lib64
-
+if [[ -z "${sha1// }" ]]; then
+    rm -r lib64
+fi
 echo "running ctest"
 ctest --output-on-failure
 popd
@@ -43,10 +47,12 @@ python -m pip install -r test_requirements.txt
 
 echo "running pytest"
 #run in a new folder so that we dont load the other python code from the source, but rather run against komodo
-rm -rf tmptest
-mkdir tmptest
-cp -r python/tests tmptest/tests
-pushd tmptest
+if [[ -z "${sha1// }" ]]; then
+    rm -rf tmptest
+    mkdir tmptest
+    cp -r python/tests tmptest/tests
+    pushd tmptest
+fi
 
 export ECL_SKIP_SIGNAL=ON
 python -m pytest \
@@ -56,4 +62,3 @@ python -m pytest \
  --ignore="tests/res/enkf/test_workflow_list.py"\
  --ignore="tests/res/enkf/test_hook_manager.py"\
  --ignore="tests/legacy"
-popd
